% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyFilt.R
\name{applyFilt}
\alias{applyFilt}
\alias{applyFilt}
\alias{applyFilt.imdFilt}
\title{Title}
\usage{
applyFilt(filter_object, rRNAdata, ...)

\method{applyFilt}{imdFilt}(filter_object, rRNAdata, min_nonmiss = 2)
}
\arguments{
\item{rRNAdata}{}
}
\description{
Title
}
