% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_data.R
\name{shift_data}
\alias{shift_data}
\title{Add a shift for log transformations}
\usage{
shift_data(omicsData, shift = 0.5)
}
\arguments{
\item{shift}{Numeric value to use for the data shift.}
}
\value{
omicsData object with the e_data abundances shifted.
}
\description{
Adds a defined amount to all values in a dataset to correct for zeroes in a dataset
}
\details{
This function shifts an rRNA dataset before log transforming, so that all 0's are removed.
}
\author{
Allison Thompson
}
