% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{Apply a transformation to rRNA data}
\usage{
transform_data(rRNA_obj, method = "clr", shift = 0.5, basis_ind = NULL,
  sampleMC = FALSE, ...)
}
\arguments{
\item{method}{}
}
\description{
Apply a transformation to rRNA data
}
